Used the suspense loader for i18n so we dont get translation pop in. Updated readme to show what this project is about now. Added a few TODO's in the readme. Added the prettier import organise plugin as the eslint one didn't play nice with importing whole files.
