+++ .eslintrc.json
@@ -66,7 +66,7 @@
     "import/newline-after-import": "warn",
     "import/no-named-as-default": "off",
     "import/no-named-as-default-member": "off",
-    "import/order": ["warn", { "newlines-between": "always", "groups": ["builtin", "external"] }],
+    "import/order": "off", // using prettier plugin
     "prettier/prettier": "error",
     "react/prop-types": "off",
     "react/self-closing-comp": "warn",
