+++ .vscode/settings.json
@@ -10,6 +10,7 @@
     "fontawesome",
     "fortawesome",
     "janky",
+    "languagedetector",
     "memoized",
     "normalise",
     "precached",
