+++ package.json
@@ -56,6 +56,7 @@
         "eslint-plugin-testing-library": "^3.10.1",
         "jest-dom": "^4.0.0",
         "prettier": "^2.2.1",
+        "prettier-plugin-organize-imports": "^1.1.1",
         "ts-node": "^9.1.1"
     },
     "browserslist": [
