+++ src/App.test.tsx
@@ -1,6 +1,5 @@
 import { render, screen } from '@testing-library/react';
 import React from 'react';
-
 import App from './App';
 
 it('renders without crashing', () => {
