+++ src/components/error-fallback/ErrorFallback.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-
 import styles from './ErrorFallback.module.scss';
 
 /** An error fallback component for Sentry. This takes up the whole page so only use it when that is intended */
