+++ src/components/nav-menu/NavMenu.test.tsx
@@ -1,6 +1,5 @@
-import React from 'react';
 import { render, screen } from '@testing-library/react';
-
+import React from 'react';
 import NavMenu from './NavMenu';
 
 describe('NavMenu', () => {
