+++ src/components/nav-menu/sub-menus/InfoMenu.test.tsx
@@ -1,9 +1,8 @@
-import React from 'react';
-import { render, screen, fireEvent } from '@testing-library/react';
+import { fireEvent, render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-
-import InfoMenu from './InfoMenu';
+import React from 'react';
 import AppContext, { systemSize } from '../../../context/SystemContext';
+import InfoMenu from './InfoMenu';
 
 describe('InfoMenu', () => {
   test('it renders', () => {
