+++ src/components/nav-menu/sub-menus/PointOfInterestButton.test.tsx
@@ -1,7 +1,6 @@
-import React from 'react';
-import { render, screen, fireEvent } from '@testing-library/react';
+import { fireEvent, render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-
+import React from 'react';
 import PointOfInterestButton from './PointOfInterestButton';
 
 describe('PointOfInterestButton', () => {
