+++ src/components/the-system/ the-belt/BeltLayer.test.tsx
@@ -1,6 +1,5 @@
 import { render, screen } from '@testing-library/react';
 import React from 'react';
-
 import BeltLayer from './BeltLayer';
 
 describe('BeltLayer', () => {
