+++ src/components/the-system/ the-belt/TheBelt.test.tsx
@@ -1,6 +1,5 @@
 import { render, screen } from '@testing-library/react';
 import React from 'react';
-
 import TheBelt from './TheBelt';
 
 describe('TheBelt', () => {
