+++ src/components/the-system/TheSystem.tsx
@@ -1,11 +1,10 @@
-import React, { useContext, FC } from 'react';
-
+import React, { FC, useContext } from 'react';
 import AppContext from '../../context/SystemContext';
+import { pointsOfInterest } from '../../PointsOfInterest';
 import TheBelt from './ the-belt/TheBelt';
+import CelestialBody from './celestial-body/CelestialBody';
 import Planet from './Planet';
 import styles from './TheSystem.module.scss';
-import { pointsOfInterest } from '../../PointsOfInterest';
-import CelestialBody from './celestial-body/CelestialBody';
 
 /**
  * The Solar System! It is memoized so this only re-renders when the context is changed.
