+++ src/components/the-system/celestial-body/CelestialBody.test.tsx
@@ -1,6 +1,5 @@
 import { render, screen } from '@testing-library/react';
 import React from 'react';
-
 import { pointsOfInterest } from '../../../PointsOfInterest';
 import CelestialBody from './CelestialBody';
