+++ src/i18n.ts
@@ -1,6 +1,6 @@
 import i18n from 'i18next';
-import Backend from 'i18next-http-backend';
 import LanguageDetector from 'i18next-browser-languagedetector';
+import Backend from 'i18next-http-backend';
 import { initReactI18next } from 'react-i18next';
 
 i18n
@@ -16,9 +16,6 @@ i18n
     debug: process.env.REACT_APP_DEBUG === 'ON',
     interpolation: {
       escapeValue: false,
-    },
-    react: {
-      useSuspense: false,
     },
   });
