+++ src/i18nForTests.tsx
@@ -1,6 +1,5 @@
 import i18n from 'i18next';
 import { initReactI18next } from 'react-i18next';
-
 import translations from '../public/locales/en-GB.json';
 
 i18n.use(initReactI18next).init({
@@ -9,9 +8,6 @@ i18n.use(initReactI18next).init({
   debug: !process.env.CI, // don't span the ci logs with setup logs
   interpolation: {
     escapeValue: false, // not needed for react!!
-  },
-  react: {
-    useSuspense: false,
   },
   resources: {
     en: {
