+++ src/index.tsx
@@ -1,10 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-
 import AppWrapper from './AppWrapper';
-import * as serviceWorker from './serviceWorker';
-
 import './index.css';
+import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(<AppWrapper />, document.getElementById('root'));
